@@ -1,6 +1,6 @@
 <?php
 
-class ConceptPropertyTest extends PHPUnit_Framework_TestCase
+class ConceptPropertyTest extends PHPUnit\Framework\TestCase
 {
   private $model; 
 
