@@ -1,6 +1,6 @@
 <?php
 
-class RequestTest extends PHPUnit_Framework_TestCase
+class RequestTest extends PHPUnit\Framework\TestCase
 {
   private $model;
   private $request;
